feat: complete visual redesign, gameplay feedback, and high-precision HUD
- Redesign gameplay background with a realistic procedural sunset theme, including snow-capped mountains, layered forest, pulsing sun, and drifting clouds.
- Implement a detailed black cat paw with realistic fur highlights, pink pads, and retractable claws.
- Replace game target with an animated bee featuring fast-flapping wings and enhanced white trail effects.
- Add haptic feedback (vibration) and festive tones (ToneGenerator) upon successful bee capture.
- Enhance UI/HUD with bold white counters and scale-up "pop" animations triggered by value changes.
- Upgrade timer to high-precision (minutes:seconds:hundredths) and modified GameState/ViewModel to support Float-based time tracking.
- Add a persistent rainbow-colored dashed trajectory line for real-time aiming guidance.
- Update system UI theme to match the sunset color palette, including a deep blue status bar with light icons.
- Optimize rendering by using graphicsLayer for HUD animations and refined procedural drawing paths.
- Added VIBRATE permission to AndroidManifest.

+++ app/src/main/java/com/mafg/mafg/mafg/mafg/catsadventures/domain/model/GameState.kt
@@ -11,6 +11,6 @@ data class GameState(
     val currentChapter: Int = 1,
     val isLoading: Boolean = false,
     val score: Int = 10,
-    val timeLeft: Int = 50,
+    val timeLeft: Float = 50f, // Cambiado a Float para milisegundos
     val isRainbowActive: Boolean = false
 )
